+++ kernel/core/src/runtime/c_runtime_stubs.cpp
@@ -1,4 +1,4 @@
-﻿#include <ntddk.h>
+﻿#include "fclmusa/platform.h"
 #include <locale.h>
 #include <__msvc_xlocinfo_types.hpp>
 #include <xlocinfo>